server/pyproject.toml
[tool.poetry]
name = "medical-project"
version = "0.1.0"
description = ""
authors = ["George Yakovlev <georgy.yakovlev@dualbootpartners.com>"]
readme = "README.md"
packages = [{include = "medical_project"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
pillow = "^10.1.0"
drf-yasg = "^1.21.7"
django-filter = "^23.5"
drf-writable-nested = "^0.7.0"
django-storages = "^1.14.2"
black = "^23.12.0"
isort = "^5.13.2"
factory-boy = "^3.3.0"
pytest-django = "^4.7.0"
faker = "^21.0.0"
django-rest-authtoken = "^2.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 100
multi_line_output = 3
profile = "black"

[tool.black]
line-length = 100
extend-exclude = '.git'

[tool.pytest.ini_options]
